Move providers to inside package folder
