+++ aurorae/providers/exceptions.py
