+++ tests/providers/spreadsheet/test_models.py
@@ -4,7 +4,7 @@ import pytest
 from pydantic.error_wrappers import ValidationError
 
 from aurorae.cnab240.v10_7.types import RegistrationType, RegistrationTypeEnum
-from providers.spreadsheet.models import (
+from aurorae.providers.spreadsheet.models import (
     Spreadsheet,
     SpreadsheetCompany,
     SpreadsheetEmployee,
